Moved default implementations into impl package
