+++ src/main/java/de/dhbw/swe/betterjeweled/core/GameManager.java
@@ -1,6 +1,7 @@
 package de.dhbw.swe.betterjeweled.core;
 
 import com.google.common.eventbus.*;
+import de.dhbw.swe.betterjeweled.core.*;
 import lombok.*;
 import lombok.experimental.*;
