+++ src/main/java/de/dhbw/swe/betterjeweled/impl/DefaultRegionFinder.java
@@ -1,4 +1,6 @@
-package de.dhbw.swe.betterjeweled.core;
+package de.dhbw.swe.betterjeweled.impl;
+
+import de.dhbw.swe.betterjeweled.core.*;
 
 import java.util.*;
 import java.util.function.*;
