+++ src/test/java/de/dhbw/swe/betterjeweled/impl/MergedRegionsTest.java
@@ -1,5 +1,6 @@
-package de.dhbw.swe.betterjeweled.core;
+package de.dhbw.swe.betterjeweled.impl;
 
+import de.dhbw.swe.betterjeweled.core.*;
 import org.junit.jupiter.api.Test;
 
 import java.awt.*;
